Move static slotting to separate module
