+++ lib/core-integration/src/Test/Integration/Framework/DSL.hs
@@ -151,7 +151,7 @@ import Cardano.Wallet.Api.Types
     , ApiAddress
     , ApiByronWallet
     , ApiCoinSelection
-    , ApiEpochInfo (..)
+    , ApiEpochInfo (ApiEpochInfo)
     , ApiFee
     , ApiNetworkInformation
     , ApiNetworkParameters (..)
@@ -190,6 +190,8 @@ import Cardano.Wallet.Primitive.AddressDerivation.Jormungandr
     ( generateKeyFromSeed )
 import Cardano.Wallet.Primitive.AddressDerivation.Shelley
     ( ShelleyKey )
+import Cardano.Wallet.Primitive.Slotting
+    ( SlotParameters (..), epochStartTime )
 import Cardano.Wallet.Primitive.SyncProgress
     ( SyncProgress (..) )
 import Cardano.Wallet.Primitive.Types
@@ -202,7 +204,6 @@ import Cardano.Wallet.Primitive.Types
     , HistogramBar (..)
     , PoolId (..)
     , SlotLength (..)
-    , SlotParameters (..)
     , SortOrder (..)
     , TxIn (..)
     , TxOut (..)
@@ -1715,10 +1716,10 @@ mkEpochInfo
     -> SlotParameters
     -- ^ Blockchain slot parameters
     -> ApiEpochInfo
-mkEpochInfo epochNo sp = ApiEpochInfo
-    { epochNumber = ApiT epochNo
-    , epochStartTime = W.epochStartTime sp epochNo
-    }
+mkEpochInfo epochNo sp =
+    ApiEpochInfo
+        (ApiT epochNo)
+        (epochStartTime sp epochNo)
 
 -- | Wallet not delegating and not about to join any stake pool.
 notDelegating
