+++ lib/core/src/Cardano/Wallet/Api/Server.hs
@@ -265,8 +265,6 @@ import Cardano.Wallet.Primitive.Types
     , UnsignedTx (..)
     , WalletId (..)
     , WalletMetadata (..)
-    , slotAt
-    , slotMinBound
     )
 import Cardano.Wallet.Registry
     ( HasWorkerCtx (..)
@@ -385,6 +383,7 @@ import qualified Cardano.Wallet as W
 import qualified Cardano.Wallet.Network as NW
 import qualified Cardano.Wallet.Primitive.AddressDerivation.Byron as Byron
 import qualified Cardano.Wallet.Primitive.AddressDerivation.Icarus as Icarus
+import qualified Cardano.Wallet.Primitive.Slotting as W
 import qualified Cardano.Wallet.Primitive.Types as W
 import qualified Cardano.Wallet.Registry as Registry
 import qualified Data.Aeson as Aeson
@@ -1418,7 +1417,7 @@ getNetworkInformation
 getNetworkInformation (_block0, np, st) nl = do
     now <- liftIO getCurrentTime
     nodeTip <- liftHandler (NW.currentNodeTip nl)
-    let ntrkTip = fromMaybe slotMinBound (slotAt sp now)
+    let ntrkTip = fromMaybe W.slotMinBound (W.slotAt sp now)
     let nextEpochNo = unsafeEpochSucc (ntrkTip ^. #epochNumber)
     pure $ ApiNetworkInformation
         { syncProgress =
