+++ lib/core/src/Cardano/Wallet/DB/Sqlite.hs
@@ -101,6 +101,8 @@ import Cardano.Wallet.Primitive.AddressDerivation
     , SoftDerivation (..)
     , WalletKey (..)
     )
+import Cardano.Wallet.Primitive.Slotting
+    ( SlotParameters, slotParams, slotStartTime )
 import Control.Concurrent.MVar
     ( modifyMVar, modifyMVar_, newMVar, readMVar )
 import Control.Exception
@@ -1038,7 +1040,7 @@ mkTxMetaEntity wid txid meta = TxMeta
 -- note: TxIn records must already be sorted by order
 -- and TxOut records must already be sorted by index
 txHistoryFromEntity
-    :: W.SlotParameters
+    :: SlotParameters
     -> W.BlockHeader
     -> [TxMeta]
     -> [(TxIn, Maybe TxOut)]
@@ -1063,7 +1065,7 @@ txHistoryFromEntity sp tip metas ins outs ws =
         , W.txInfoDepth =
             Quantity $ fromIntegral $ if tipH > txH then tipH - txH else 0
         , W.txInfoTime =
-            W.slotStartTime sp (meta ^. #slotId)
+            slotStartTime sp (meta ^. #slotId)
         }
       where
         txH  = getQuantity (meta ^. #blockHeight)
@@ -1330,7 +1332,7 @@ selectTxHistory wid minWithdrawal order conditions = do
 
             let wal = checkpointFromEntity cp [] ()
             let tip = W.currentTip wal
-            let slp = W.slotParams $ W.blockchainParameters wal
+            let slp = slotParams $ W.blockchainParameters wal
 
             return $ txHistoryFromEntity slp tip metas ins outs ws
   where
