+++ lib/core/src/Cardano/Wallet/DB/Sqlite/Types.hs
@@ -24,6 +24,8 @@ import Cardano.Wallet.Primitive.AddressDerivation
     ( AccountingStyle (..), Passphrase (..), PassphraseScheme (..) )
 import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( AddressPoolGap (..), getAddressPoolGap, mkAddressPoolGap )
+import Cardano.Wallet.Primitive.Slotting
+    ( flatSlot, fromFlatSlot )
 import Cardano.Wallet.Primitive.Types
     ( Address (..)
     , ChimericAccount (..)
@@ -43,8 +45,6 @@ import Cardano.Wallet.Primitive.Types
     , StakePoolTicker
     , TxStatus (..)
     , WalletId (..)
-    , flatSlot
-    , fromFlatSlot
     , isValidCoin
     , unsafeEpochNo
     )
