+++ lib/core/test/shared/Cardano/Wallet/DummyTarget/Primitive/Types.hs
@@ -17,6 +17,8 @@ module Cardano.Wallet.DummyTarget.Primitive.Types
 
 import Prelude
 
+import Cardano.Wallet.Primitive.Slotting
+    ( slotMinBound )
 import Cardano.Wallet.Primitive.Types
     ( ActiveSlotCoefficient (..)
     , Block (..)
@@ -35,7 +37,6 @@ import Cardano.Wallet.Primitive.Types
     , TxIn (..)
     , TxOut (..)
     , TxParameters (..)
-    , slotMinBound
     )
 import Crypto.Hash
     ( Blake2b_256, hash )
