+++ lib/core/test/unit/Cardano/WalletSpec.hs
@@ -64,6 +64,8 @@ import Cardano.Wallet.Primitive.CoinSelection
     ( CoinSelection, feeBalance )
 import Cardano.Wallet.Primitive.Fee
     ( Fee (..) )
+import Cardano.Wallet.Primitive.Slotting
+    ( flatSlot, unsafeEpochNo )
 import Cardano.Wallet.Primitive.SyncProgress
     ( SyncTolerance (..) )
 import Cardano.Wallet.Primitive.Types
@@ -94,9 +96,7 @@ import Cardano.Wallet.Primitive.Types
     , WalletId (..)
     , WalletMetadata (..)
     , WalletName (..)
-    , flatSlot
     , txId
-    , unsafeEpochNo
     )
 import Cardano.Wallet.Transaction
     ( ErrMkTx (..), TransactionLayer (..) )
